Introduce ExternalDomainCtTypeFactory
